shift overlap checking to its own file, add test case
