+++ pkg/service/accesssvc/add_review_test.go
@@ -7,9 +7,7 @@ import (
 
 	"github.com/benbjohnson/clock"
 	"github.com/common-fate/granted-approvals/pkg/access"
-	"github.com/common-fate/granted-approvals/pkg/rule"
 	"github.com/common-fate/granted-approvals/pkg/storage"
-	"github.com/common-fate/granted-approvals/pkg/types"
 
 	"github.com/common-fate/ddb/ddbmock"
 	"github.com/common-fate/granted-approvals/pkg/service/accesssvc/mocks"
@@ -17,121 +15,6 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestOverlapsExistingGrant(t *testing.T) {
-	type testcase struct {
-		name               string
-		accessRequest      access.Request
-		upcomingRequests   []access.Request
-		currentRequestRule rule.AccessRule
-		allRules           []rule.AccessRule
-		want               bool
-		clock              clock.Clock
-	}
-	clk := clock.NewMock()
-	inOneMinute := clk.Now().Add(time.Minute)
-
-	now := clk.Now()
-
-	//some requests premade with specific timings
-	activeRequest := access.Request{Status: access.Status(types.RequestStatusAPPROVED), Grant: &access.Grant{Status: "ACTIVE"}, Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 2}}
-	cancelledRequest := access.Request{Status: access.Status(types.RequestStatusCANCELLED), RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 2}, Rule: "rule_a"}
-	revokedRequest := access.Request{Status: access.Status(types.RequestStatusCANCELLED), Grant: &access.Grant{Status: "REVOKED"}, RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 2}, Rule: "rule_a"}
-
-	args1 := make(map[string]string)
-	args1["1"] = "arg1"
-	args1["2"] = "arg2"
-
-	args2 := make(map[string]string)
-	args2["a"] = "argA"
-	args2["b"] = "argB"
-
-	testcases := []testcase{
-		{
-			name:               "no existing grants",
-			accessRequest:      access.Request{ID: "123"},
-			upcomingRequests:   []access.Request{},
-			currentRequestRule: rule.AccessRule{Target: rule.Target{}},
-			allRules:           []rule.AccessRule{},
-			clock:              clk,
-			want:               false,
-		},
-		{
-			name:               "different provider passes",
-			accessRequest:      access.Request{ID: "abc", Rule: "rule_b"},
-			upcomingRequests:   []access.Request{{ID: "def", Rule: "rule_a"}},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a"}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_b"}}},
-			clock:              clk,
-			want:               false,
-		},
-		{
-			name:               "request overlaps current active request fails",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{activeRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a"}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a"}}},
-			clock:              clk,
-			want:               true,
-		},
-		{
-			name:               "scheduled request overlaps current active request fails",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &inOneMinute, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{activeRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a"}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a"}}},
-			clock:              clk,
-			want:               true,
-		},
-
-		{
-			name:               "same rule different arguments should succeed",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &inOneMinute, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{activeRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args2}}},
-			clock:              clk,
-			want:               false,
-		},
-		{
-			name:               "same rule same arguments should fail",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &inOneMinute, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{activeRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}}},
-			clock:              clk,
-			want:               true,
-		},
-		{
-			name:               "same rule same arguments on expired request should pass",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{cancelledRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}}},
-			clock:              clk,
-			want:               false,
-		},
-		{
-			name:               "same rule same arguments on revoked request should pass",
-			accessRequest:      access.Request{ID: "123", Rule: "rule_a", RequestedTiming: access.Timing{StartTime: &now, Duration: time.Minute * 5}},
-			upcomingRequests:   []access.Request{revokedRequest},
-			currentRequestRule: rule.AccessRule{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}},
-			allRules:           []rule.AccessRule{{ID: "rule_a", Target: rule.Target{ProviderID: "prov_a", With: args1}}},
-			clock:              clk,
-			want:               false,
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			got, err := overlapsExistingGrantCheck(tc.accessRequest, tc.upcomingRequests, tc.currentRequestRule, tc.allRules, tc.clock)
-			assert.NoError(t, err)
-			assert.Equal(t, tc.want, got)
-		})
-
-	}
-
-}
 
 func TestAddReview(t *testing.T) {
 	type createGrantResponse struct {
